Feat: http inspector 적용
